requirements.txt
openai>=0.27.0
pyyaml>=5.4.1
tiktoken>=0.4.0
